[SME] WIP: prototype opaque tile types and allocation
This patch extends SME support in LLVM with target extension types [1]
for each tile size, these are:

  - target("aarch64.za.b")
  - target("aarch64.za.h")
  - target("aarch64.za.s")
  - target("aarch64.za.d")
  - target("aarch64.za.q")

The register classes are updated from being untyped, to using the opaque
types. An initial set of new intrinsics that use the opaque types are
also added for loads, stores and vector <-> tile, along with (mostly
duplicated) patterns to match these to instructions. The intrinsics have
the (temporary) prefix 'int_aarch64_sme_x'.

I've made a stab at implementing spilling and filling support for tile
registers (well, only 16-bit tiles, but the approach should scale to
other tile types). This is incomplete, but the general approach is to
emit spill/fill pseudo instructions in storeRegToStackSlot /
loadRegFromStackSlot with operands that will later be forwarded to the
actual load/store instructions, as well as a frame index containing a
pointer to the stack slot to spill to. These pseudos are expanded post
register allocation in expandPostRAPseudo, where a loop is materialized
that spills/fills the tile to/from the stack a tile slice at a time,
e.g. for a 32-bit tile:

.spill.loop
  rdsvl   x13, #1                       // SVL bytes.
  mul     x13, x13, x13                 // SVL * SVL bytes (total size of ZA in bytes).
  sub     sp, sp, x13, lsr #2           // Shift SVL * SVL bytes right by 2 to account for (smaller) 32-bit tile size and adjust SP to allocate [2].
  mov     x10, sp
  mov     w12, wzr                      // tile slice index.
  cntw    x8                            // # of tile slices in 32-bit tile.
  ptrue   p0.s
  st1w    {za0h.s[w12, 0]}, p0/z, [x10] // store tile slice.
  add     x12, x12, #1                  // increment tile slice index by 1.
  addvl   x9, x9, #1                    // adjust base (sp) by size of tile slice.
  cmp     x12, x8                       // while (tile slice index < # of tile slices).
  b.lt    .spill.loop

I initially tried to materialize the loop directly (without pseudos) in
storeRegToStackSlot / loadRegFromStackSlot but the isKnownSentinel
assert kept firing, I think because the MBBI iterator was being
incremented past the end (?). Lowering to pseudos instead and
materializing the loop in expandPostRAPseudo got further, but again I
hit this assert, and hacked around it this time by inserting a couple of
nops.

The operands for the loads/stores in expandPostRAPseudo are taken from
the spill/fill pseudos rather than creating them directly in
expandPostRAPseudo.  The reason for this is virtual registers can't be
created post regalloc, but perhaps registers could be scavenged (?).

The SP adjustment to allocate space to spill a tile is not implemented,
but the plan was to add a new stack ID (ScalableMatrix?), then in
AArch64FrameLowering make the adjustment to SP for each frame index with
this stack ID, then store the SP in the frame index.

The two key things opaque types could address are:

  - intrinsics that model the dataflow
  - register allocation

In MLIR, the former was achieved with the current intrinsics via three
ops:

  - arm_sme.get_tile_id
  - arm_sme.cast_tile_to_vector
  - arm_sme.cast_vector_tile

These are discussed along with the design in [3]. Register allocation is
partially implemented in MLIR but spilling / filling is not supported,
an error will occur when running out of tiles. I've began updating the
ArmSME dialect in MLIR to use the new set of intrinsics. A new op
'arm_sme.get_tile' is added that returns rank-2 scalable vector types
correlating with SME tiles. This is converted to undef when lowering to
LLVM. Type conversions to the opaque types are added. There are still
ops using the casts, notably 'arm_sme.zero'.

This goal of this prototype was foremost to explore what is required for
spilling / filling in the backend (dataflow is intertwined with this) vs
MLIR. I think this approach is feasible with more work, but from the
MLIR side as we increase support and use more intrinsics there is a
significant cost in supporting the new intrinsics and related patterns.
We'll be exploring spilling / filling at a higher-level in MLIR soon.

Other changes:
- Updated the load instruction to take a tile as input that's tied to
  the output. Since it's only loading a slice, it's currently not modelled
  correctly.
- Set the alignment of the tile registers to 16-bytes. It's currently
  aligning to the minimum size of a tile.

[1] https://llvm.org/docs/LangRef.html#target-extension-type
[2] While the size of a vector of SVL bits is constant at runtime, the number
of such vectors in a tile varies based on the element size. This is summarized
in the table below.

  SVL [bits] SVLB SVLH SVLS SVLD SVLQ
         128   16    8    4    2    1
         256   32   16    8    4    2
         512   64   32   16    8    4
        1024  128   64   32   16    8
        2048  256  128   64   32   16
  RDSVL^2 >>    0    1    2    3    4  //=tile size in bytes

[3] https://discourse.llvm.org/t/an-update-on-armsme-dialect-and-proposed-next-steps/72036

Current failing tests (excl. Clang):
--------------------------------------------------------------------------------
Failed Tests (16):
  // Think patterns need updating now register classes are typed.
  LLVM :: CodeGen/AArch64/sme2-intrinsics-extract-mova.ll

  // load_tile_slice ops need updating to add passthru tile. Vector <->
  // tile ops need updating since intrinsics have changed.
  MLIR :: Dialect/ArmSME/arm-sme-to-llvm.mlir
  MLIR :: Dialect/ArmSME/vector-ops-to-llvm.mlir

  // unrealized_conversion_cast between Vector tile types and opaque
  // types. ('arm_sme.zero' needs updating').
  MLIR :: Integration/Dialect/Linalg/CPU/ArmSME/fill-2d.mlir
  MLIR :: Integration/Dialect/Linalg/CPU/ArmSME/matmul-transpose-a.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-outerproduct-f32.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-outerproduct-f64.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-transfer-read-2d.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-transfer-write-2d.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/tile_fill.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/vector-ops.mlir

  // 'storeRegToStackSlot not implemented for .S tiles!' firing because
  // of copies between MPRs.
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-load-vertical.mlir
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/test-transpose.mlir

  // 'storeRegToStackSlot not implemented for .D tiles!' firing because
  // of copies between MPRs. 'load_store_two_za_s_tiles' does work tho.
  MLIR :: Integration/Dialect/Vector/CPU/ArmSME/vector-load-store.mlir

  // Just needs updating since changes to ops.
  MLIR :: Target/LLVMIR/arm-sme-invalid.mlir